Organized files into rofication package
