+++ rofication-gui.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+import sys
+
+from rofication import Rofication
+
+if __name__ == '__main__':
+    Rofication(sys.stdout).run()
